Initializing RefCounted::_refCount to zero is way cool.
Made RefCounted::ref and unref protected, and made RefPtr<T> friends with
RefCounted.  Maybe this isn't so hot; if we ever want to push the envelope
and do ref counting manually, we are now out of luck.  On the other hand, that
just means that there are fewer ways we can fuck up.

The editor can create geometry nodes now. (still can't do anything with them,
though)


git-svn-id: file:///desktop/SourceForge%20Backup/empyrean/svnrepos/trunk@228 fea1f14b-4e95-431d-9825-d582385f3f73